Abstracted db connection

+++ types/totalfiles.go
@@ -1,5 +1,5 @@
 package types
 
 type TotalBytes struct {
-	Bytes int64   `json:"bytes"`
+	Bytes int   `json:"bytes"`
 }
